+++ src/Exceptions/GlideUrlException.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace AmpedWeb\GlideInABox\Exceptions;
+
+
+use Exception;
+
+class GlideUrlException extends Exception
+{
+
+}