+++ src/Exceptions/InvalidSizeFitException.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace AmpedWeb\GlideInABox\Exceptions;
+
+
+class InvalidSizeFitException extends InvalidSizeException
+{
+
+}