+++ src/Util/GlideUrl.php
@@ -4,12 +4,18 @@
 namespace AmpedWeb\GlideInABox\Util;
 
 
+use AmpedWeb\GlideInABox\Traits\Crop;
+use AmpedWeb\GlideInABox\Traits\Encode;
+use AmpedWeb\GlideInABox\Traits\Flip;
+use AmpedWeb\GlideInABox\Traits\Orientation;
+use AmpedWeb\GlideInABox\Traits\Size;
 use Illuminate\Support\Str;
 use League\Glide\Urls\UrlBuilder;
 use League\Glide\Urls\UrlBuilderFactory;
 
 class GlideUrl
 {
+    use Orientation, Flip, Crop, Size, Encode;
 
     /**
      * The filepath of our image being manipulated
@@ -22,6 +28,10 @@ class GlideUrl
      */
     protected $urlFactory;
 
+    /**
+     * @var array
+     */
+    protected $buildParams;
 
     /**
      * GlideUrl constructor.
@@ -29,6 +39,7 @@ class GlideUrl
     public function __construct()
     {
         $this->urlFactory = UrlBuilderFactory::create('/' . config('glideinabox.base_url') . '/', config('glideinabox.signature_key'));
+        $this->buildParams = [];
     }
 
 
@@ -46,21 +57,13 @@ class GlideUrl
     /**
      * @return mixed
      */
-    protected function parsedPath()
+    public function getParsedPath()
     {
         if (Str::startsWith($this->path, '/storage/')) {
             $this->path = Str::replaceFirst('/storage/', '', $this->path);
         }
 
         return Str::of($this->path)->replace('\\', '/');
-    }
-
-    /**
-     * @return string
-     */
-    public function getParsedPath():string
-    {
-        return $this->parsedPath();
     }
 
     /**
@@ -82,14 +85,17 @@ class GlideUrl
     }
 
     /**
+     * Start building an image URL based on a preset
+     *
      * @param       $presets
      * @param array $params
      *
-     * @return mixed
+     * @return GlideUrl
      */
-    public function preset($presets, array $params = [])
+    public function preset($presets, array $params = []): GlideUrl
     {
-        return url($this->urlFactory->getUrl($this->parsedPath(), array_merge($this->parsePresets($presets), $params)));
+        $this->buildParams = array_merge($this->parsePresets($presets), $params);
+        return $this;
     }
 
     /**
@@ -100,7 +106,43 @@ class GlideUrl
     public function custom(array $params = [])
     {
 
-        return url($this->urlFactory->getUrl($this->parsedPath(), $params));
+        return url($this->urlFactory->getUrl($this->getParsedPath(), $params));
     }
 
+    /*
+     * Fluent Interface Functions
+     */
+
+
+    /**
+     * Start building an image configuration
+     *
+     * @return $this
+     */
+    public function build(): GlideUrl
+    {
+        $this->buildParams = [];
+
+        return $this;
+    }
+
+    /**
+     * Get the URL to your image after building the configuration
+     *
+     * @return string
+     */
+    public function url(): string
+    {
+        return url($this->urlFactory->getUrl($this->getParsedPath(), $this->buildParams));
+    }
+
+    /**
+     * Get the current set of image builder parameters
+     *
+     * @return array
+     */
+    public function getParams(): array
+    {
+        return $this->buildParams;
+    }
 }