+++ CHANGELOG.md
@@ -19,6 +19,14 @@ Types of changes
 ### Added
 * .gitignore
 * PHP extension requirements
+* Orientation trait
+* Encode trait
+* Crop trait
+* Flip Trait
+* Size Trait
+
+### Changed
+* GlideUrl::preset starts an image builder instead of returning an image URL
 
 ### Fixed
 * Broken unit tests in ImageResponseTest